create the initial state in the src/index.js file
