+++ src/index.js
@@ -8,6 +8,12 @@ import rootReducer from './reducers/index';
 
 const store = createStore(rootReducer);
 
+const initialState = [
+  { id: 0, title: 'Friday 13', category: 'Horror' },
+  { id: 1, title: 'Wild-life Dangers', category: 'Learning' },
+  { id: 2, title: 'Toonami', category: 'Kids' },
+];
+
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
@@ -20,3 +26,5 @@ ReactDOM.render(
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+
+export default initialState;
