--- src/reducers/initialState.js
@@ -1,7 +1,0 @@
-const initialState = [
-  { id: 0, title: 'Friday 13', category: 'Horror' },
-  { id: 1, title: 'Wild-life Dangers', category: 'Learning' },
-  { id: 2, title: 'Toonami', category: 'Kids' },
-];
-
-export default initialState;
